# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VendorBridge Procurement ERP API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  totalVendors: number;
  activeRfqs: number;
  pendingApprovals: number;
  openPurchaseOrders: number;
  unpaidInvoices: number;
  totalSpendMtd: number;
  savingsPercent: number;
  onTimeDeliveryRate: number;
}

export interface MonthlySpend {
  month: string;
  spend: number;
  budget: number;
}

export interface CategorySpend {
  category: string;
  amount: number;
  percentage: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  timestamp: string;
  status: string;
}

export interface Vendor {
  id: number;
  name: string;
  category: string;
  contactName: string;
  email: string;
  phone: string;
  /** @nullable */
  address?: string | null;
  status: string;
  rating: number;
  totalOrders: number;
  totalSpend: number;
  createdAt: string;
}

export interface VendorInput {
  name: string;
  category: string;
  contactName: string;
  email: string;
  phone: string;
  address?: string;
}

export interface VendorUpdate {
  name?: string;
  category?: string;
  contactName?: string;
  email?: string;
  phone?: string;
  address?: string;
  status?: string;
  rating?: number;
}

export interface Rfq {
  id: number;
  title: string;
  description: string;
  category: string;
  status: string;
  deadline: string;
  vendorIds: number[];
  estimatedValue: number;
  quotationCount?: number;
  createdAt: string;
}

export interface RfqInput {
  title: string;
  description: string;
  category: string;
  deadline: string;
  vendorIds: number[];
  estimatedValue: number;
}

export interface RfqUpdate {
  title?: string;
  description?: string;
  status?: string;
  deadline?: string;
}

export interface Quotation {
  id: number;
  rfqId: number;
  rfqTitle?: string;
  vendorId: number;
  vendorName: string;
  unitPrice: number;
  totalPrice: number;
  deliveryDays: number;
  paymentTerms: string;
  /** @nullable */
  notes?: string | null;
  status: string;
  submittedAt: string;
}

export interface QuotationInput {
  rfqId: number;
  vendorId: number;
  unitPrice: number;
  totalPrice: number;
  deliveryDays: number;
  paymentTerms: string;
  notes?: string;
}

export interface QuotationUpdate {
  status?: string;
  notes?: string;
}

export interface Approval {
  id: number;
  type: string;
  referenceId: number;
  referenceTitle: string;
  requestedBy: string;
  /** @nullable */
  approvedBy?: string | null;
  amount: number;
  status: string;
  priority: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  resolvedAt?: string | null;
}

export interface ApprovalAction {
  approverName: string;
  notes?: string;
}

export interface LineItem {
  description: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export interface PurchaseOrder {
  id: number;
  poNumber: string;
  vendorId: number;
  vendorName: string;
  items: LineItem[];
  totalAmount: number;
  status: string;
  deliveryDate: string;
  /** @nullable */
  deliveredAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PurchaseOrderInput {
  vendorId: number;
  items: LineItem[];
  deliveryDate: string;
  notes?: string;
}

export interface PurchaseOrderUpdate {
  status?: string;
  deliveredAt?: string;
  notes?: string;
}

export interface Invoice {
  id: number;
  invoiceNumber: string;
  vendorId: number;
  vendorName: string;
  purchaseOrderId: number;
  poNumber: string;
  amount: number;
  status: string;
  dueDate: string;
  issuedAt: string;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface InvoiceInput {
  vendorId: number;
  purchaseOrderId: number;
  amount: number;
  dueDate: string;
  issuedAt: string;
  notes?: string;
}

export interface InvoiceUpdate {
  status?: string;
  paidAt?: string;
  notes?: string;
}

export interface VendorPerformance {
  vendorId: number;
  vendorName: string;
  onTimeRate: number;
  qualityScore: number;
  totalOrders: number;
  totalSpend: number;
}

export interface ProcurementPipeline {
  draftRfqs: number;
  activeRfqs: number;
  pendingQuotations: number;
  pendingApprovals: number;
  activePOs: number;
  pendingInvoices: number;
}